+++ CHANGELOG.md
@@ -82,6 +82,14 @@ All notable changes for each version of this project will be documented in this 
 - `IgxDateTimeEditor`, `IgxMask`, `IgxDatePicker`, `IgxTimePicker`, `IgxDateRangePicker`
     - Added IME input support. When typing in an Asian language input, the control will display input method compositions and candidate lists directly in the control’s editing area, and immediately re-flow surrounding text as the composition ends.
 
+- `IgxCombo`
+    - Added `selectionMode` `@Input`. Allows to specify whether the selection mode should be 'multiple' (default, current behavior) or 'single' (allowing only one item to be selected).
+    ```html
+    <igx-combo #single [data]="myData" selectionMode="single">
+        ...
+    </igx-combo>
+    ``` 
+
 ### General
 - `IgxGridComponent`
     - The following properties are deprecated: